Use a maybe CE to simplify null navigation
Copy the MaybeBuilder from FSharpx.Extras and augment it with support for null references and Nullable values.

Only the builder itself is used so a source code copy was preferred to bringing in an entire nuget package.

Also fixes the build output not copying FSharp.Core.dll, a necessary dependency.